refactor(ci): Combine CI+Master for Controller and Service
In the next commit we'll be subclassing these for each type of
CI; it makes more sense to split the files along the lines of
CI type than to have separate classes for the CI and the Master.
