+++ halyard-config/src/main/java/com/netflix/spinnaker/halyard/config/services/v1/ci/CiService.java
@@ -19,10 +19,12 @@ package com.netflix.spinnaker.halyard.config.services.v1.ci;
 import com.netflix.spinnaker.halyard.config.error.v1.ConfigNotFoundException;
 import com.netflix.spinnaker.halyard.config.error.v1.IllegalConfigException;
 import com.netflix.spinnaker.halyard.config.model.v1.node.Ci;
+import com.netflix.spinnaker.halyard.config.model.v1.node.Master;
 import com.netflix.spinnaker.halyard.config.model.v1.node.NodeFilter;
 import com.netflix.spinnaker.halyard.config.problem.v1.ConfigProblemBuilder;
 import com.netflix.spinnaker.halyard.config.services.v1.LookupService;
 import com.netflix.spinnaker.halyard.config.services.v1.ValidateService;
+import com.netflix.spinnaker.halyard.core.error.v1.HalException;
 import com.netflix.spinnaker.halyard.core.problem.v1.Problem.Severity;
 import com.netflix.spinnaker.halyard.core.problem.v1.ProblemSet;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -93,6 +95,86 @@ public class CiService {
         .withAnyCi()
         .withAnyAccount();
 
+    return validateService.validateMatchingFilter(filter);
+  }
+
+  public List<Master> getAllMasters(String deploymentName, String ciName) {
+    NodeFilter filter = new NodeFilter().setDeployment(deploymentName).setCi(ciName).withAnyMaster();
+
+    List<Master> matchingMasters = lookupService.getMatchingNodesOfType(filter, Master.class);
+
+    if (matchingMasters.size() == 0) {
+      throw new ConfigNotFoundException(
+          new ConfigProblemBuilder(Severity.FATAL, "No masters could be found").build());
+    } else {
+      return matchingMasters;
+    }
+  }
+
+  private Master getMaster(NodeFilter filter, String masterName) {
+    List<Master> matchingMasters = lookupService.getMatchingNodesOfType(filter, Master.class);
+
+    switch (matchingMasters.size()) {
+      case 0:
+        throw new ConfigNotFoundException(new ConfigProblemBuilder(
+            Severity.FATAL, "No master with name \"" + masterName + "\" was found")
+            .setRemediation("Check if this master was defined in another Continuous Integration service, or create a new one").build());
+      case 1:
+        return matchingMasters.get(0);
+      default:
+        throw new IllegalConfigException(new ConfigProblemBuilder(
+            Severity.FATAL, "More than one master named \"" + masterName + "\" was found")
+            .setRemediation("Manually delete/rename duplicate masters with name \"" + masterName + "\" in your halconfig file").build());
+    }
+  }
+
+  public Master getCiMaster(String deploymentName, String ciName, String masterName) {
+    NodeFilter filter = new NodeFilter().setDeployment(deploymentName).setCi(ciName).setMaster(masterName);
+    return getMaster(filter, masterName);
+  }
+
+  public Master getAnyCiMaster(String deploymentName, String masterName) {
+    NodeFilter filter = new NodeFilter().setDeployment(deploymentName).withAnyCi().setMaster(masterName);
+    return getMaster(filter, masterName);
+  }
+
+  public void setMaster(String deploymentName, String ciName, String masterName, Master newMaster) {
+    Ci ci = getCi(deploymentName, ciName);
+
+    for (int i = 0; i < ci.getMasters().size(); i++) {
+      Master master = (Master) ci.getMasters().get(i);
+      if (master.getNodeName().equals(masterName)) {
+        ci.getMasters().set(i, newMaster);
+        return;
+      }
+    }
+
+    throw new HalException(new ConfigProblemBuilder(Severity.FATAL, "Master \"" + masterName + "\" wasn't found").build());
+  }
+
+  public void deleteMaster(String deploymentName, String ciName, String masterName) {
+    Ci ci = getCi(deploymentName, ciName);
+    boolean removed = ci.getMasters().removeIf(master -> ((Master) master).getName().equals(masterName));
+
+    if (!removed) {
+      throw new HalException(
+          new ConfigProblemBuilder(Severity.FATAL, "Master \"" + masterName + "\" wasn't found")
+              .build());
+    }
+  }
+
+  public void addMaster(String deploymentName, String ciName, Master newMaster) {
+    Ci ci = getCi(deploymentName, ciName);
+    ci.getMasters().add(newMaster);
+  }
+
+  public ProblemSet validateMaster(String deploymentName, String ciName, String masterName) {
+    NodeFilter filter = new NodeFilter().setDeployment(deploymentName).setCi(ciName).setMaster(masterName);
+    return validateService.validateMatchingFilter(filter);
+  }
+
+  public ProblemSet validateAllMasters(String deploymentName, String ciName) {
+    NodeFilter filter = new NodeFilter().setDeployment(deploymentName).setCi(ciName).withAnyMaster();
     return validateService.validateMatchingFilter(filter);
   }
 }
